Don't enable all systemd flags for OpenRC
I know the comment says it's for future flexibility, but if switching
from OpenRC to systemd, you're going to have to rebuild your kernel
anyway. This moves the few systemd-specific kernel configs into a block
that's only run if targetting systemd. The corresponding OpenRC block
doesn't require any specific configs that I could determine, but left
empty in case anything needs to be added in the future. I've built an
OpenRC-based system with a clean kernel config with this and confirmed
it works.